requirements.txt
PyYAML
django-yamlfield
termcolor
